@@ -6,41 +6,39 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server",
-    "build": "rimraf dist && webpack --config ./webpack.production.config.js --progress --profile --colors"
+    "build": "rimraf dist && webpack --config ./webpack.production.config.js -p --progress --profile --colors"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "babel-polyfill": "^6.16.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^3.0.0",
-    "react-router-redux": "^4.0.6",
+    "babel-polyfill": "^6.23.0",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^4.4.8",
+    "react-router-dom": "^4.1.1",
     "redux": "^3.6.0",
-    "redux-thunk": "^2.1.0"
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-core": "^6.18.0",
-    "babel-loader": "^6.2.7",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "css-loader": "^0.25.0",
-    "eslint": "^3.10.2",
-    "eslint-plugin-react": "^6.7.1",
-    "extract-text-webpack-plugin": "^2.0.0-rc.2",
-    "html-webpack-plugin": "^2.24.0",
-    "less": "^2.7.1",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^6.10.3",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "html-webpack-plugin": "^2.28.0",
+    "less": "^2.7.2",
     "less-loader": "^2.2.3",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "redux-devtools": "^3.3.1",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.1.1",
-    "rimraf": "^2.5.4",
-    "stats-webpack-plugin": "^0.4.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.3.0"
+    "postcss-loader": "^2.0.5",
+    "react-hot-loader": "^3.0.0-beta.7",
+    "rimraf": "^2.6.1",
+    "stats-webpack-plugin": "^0.4.3",
+    "style-loader": "^0.13.2",
+    "webpack": "^2.6.1",
+    "webpack-dev-server": "^2.4.5"
   }
 }
