@@ -1,5 +1,11 @@
-if (process.env.NODE_ENV === 'production') {
-    module.exports = require('./create.prod');
-} else {
-    module.exports = require('./create.dev');
+import { createStore as _createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import reducer from './modules/reducer';
+
+export default function createStore() {
+    const middleware = [thunk];
+    const finalCreateStore = applyMiddleware(...middleware)(_createStore);
+    const store = finalCreateStore(reducer);
+
+    return store;
 }
