@@ -1,9 +1,8 @@
 import React from 'react';
-import {hashHistory} from 'react-router';
-const About = () =>
+const About = ({history}) =>
     <div>
         <p>xxxholic</p>
-        <button onClick={()=>hashHistory.goBack()}>go back</button>
+        <button onClick={()=>history.goBack()}>go back</button>
     </div>;
 
 
