@@ -1,22 +1,30 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import { withRouter } from 'react-router-dom';
 import {
     increment,
     incrementAsync
 } from '../../redux/modules/home';
-import { hashHistory } from 'react-router';
+
+import './home.less';
 
 class Home extends Component {
+   
     render() {
-        const {count, increment, incrementAsync} = this.props;
+        const {count, increment, incrementAsync,history} = this.props;
 
         return (
             <div>
                 <h1 className="home-count">{count}</h1>
+                <div className="tab">
+                    <div>1</div>
+                    <div>2</div>
+                    <div>3</div>
+                </div>
                 <button onClick={increment}>Increment</button>
                 <button onClick={incrementAsync}>incrementAsync</button>
-                <button onClick={() => hashHistory.push('/about')}>goAbout</button>
+                <button onClick={() => history.push('/about')}>goAbout</button>
             </div>
         );
     }
@@ -35,7 +43,7 @@ const mapDispatchToProps = (dispatch) => bindActionCreators(
     }, dispatch);
 
 
-export default connect(
+export default withRouter(connect(
     mapStateToProps,
     mapDispatchToProps
-)(Home);
+)(Home));
